Tested and improved scripts and documentation (#11)
* Fixed paths and done first steps for documentation overhaul

* Added dependency installation step and fixed shebangs and set executable bits on python files in mkmapgrids/

* Tested all scripts succesfully up to and including surface file creation

* Moved to Stages/2024

The files created with mkmapgrid/*.py, using this newer environment, yield
field values identical to when using Stages/2023.

* Made ICON_SCRIP.py simpler to use by adding command line arguments

* Cleaned up *Creation of gridfile* section of the documentation (README.md)
and made it consistent with these code changes.

* Removed hard-coded ifile and ofile

Resolves: #8

* Adjusted creation of domain files for Stages/2024

- Stages/2024 uses GCC and gfortran apparently needs explicit include of $INC_NETCDF
- Use standard Fortran functions for getting command-line arguments

https://github.com/ESMCI/cime/commit/a608d9930fbc0eda44786645bd76be6520487f3c

* Simplified argument parsing with default arguments

* Changed name to "generator" in accordance with renamed repo (#9)

- removed spurious spaces before EOL, and
- applied one-sentence per line convention for changed paragraphs.

* Clarified and completed NCL instructions and where to get source grid file

As long as there are no other clear and working instructions to create the
SCRIP file from the rotated curvilinear grid, the NCL instructions have to be
clear and complete to reproduce the files at all and later to compare with the
replaced method that should replace the NCL script!

* Added Perl module

This version, as opposed to the locally installed one, includes required Perl
packages.

* Made modules (Stages/2024), paths and variables more consistent

- Added explanation about the pan-European domain
- Made paths and filenames internally consistent
- Fixed the naming of the environment files
- Use Stages/2024 and load modules at the start to consistenly use throughout
  the static file generator
- Give choice between GCC and Intel, with provision that the Intel compiler
  needs to be used for mkmapdata/runscript_mkmapdata.sh
- Don't use non-existing paths (and be mostly independent of JSC machines)

* Use new Intel Fortran compiler binary ifx

* Bug fixes, typos and added software modules

- Trivial fixes and typos
- Rely on hashbang
- Load NCO and CDO, because needed for forcing data

* Use shell variable instead of string

* Use relative paths to files generated by workflow

* For each step, point to directory with relevant scripts

Minor improvements include extended explanations and a typo in a module version.

Ignore binary objects and mod files.

* Create the surface file with constant values

For now the strategy is to create a simple file and then replace the variables.
In the future we might adjust the procedure and feed the correct data directly
to mksurfdata.pl.

* Ignore netCDF, swap and log files

* Updated CSMDATA path to correct location

A full copy of the UCAR CLM rawdata is now in the DETECT CentralDB.

Resolves: #12 #5

* Referencing the EUR-R13B05 grid for the ICON grid

There are also working suggestions for the rectilinear and curvilinear case.

* By default use the icosahedral grid in this workflow

- Point to the EUR-R13B05 grid also in the mkmapdata script
- Path to mappingdata is now on largedata below CentralDB
- The sbatch directive --constraint is needed to give the job access

* Fixed the CSMDATA path in the README

`/lnd/clm2/rawdata` is added automatically.

* Creation of constant fields works

The script has a problem with the UCAR input data.

Some variables, at least SOIL_COLOR, will not be constant for unknown reasons.
This is essentially not a problem, because the variable will be overwritten.

* Short guiding comment about grid type to use

This will be more extensively addressed in a different merge request where we
also make the naming of grids, scripts and netCDF output files more consistent
and clear (#20).